zimmer-backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
python-dotenv==1.0.0
pydantic==2.5.0
pydantic[email]==2.5.0
python-multipart==0.0.6
openai==0.28.1
bcrypt==4.1.2
pyjwt==2.8.0
alembic==1.13.1
httpx==0.25.2
requests==2.31.0
APScheduler==3.10.4 